+++ src/wayland/output/mod.rs
@@ -1,19 +1,28 @@
 //! Output advertising capabilities
 //!
-//! This module provides a type helping you to handle the advertising
-//! of your compositor's output and their capabilities to your client,
-//! as well as mapping your clients output request to your physical
-//! outputs.
+//! This module provides a type helping you to abstract over various
+//! properties, that make up an [`Output`] of your compositor.
+//!
+//! Additionally the [`Output`] may handle advertising of your compositor's
+//! output and their capabilities to your client, as well as mapping your
+//! clients output request to your physical outputs.
 //!
 //! # How to use it
 //!
-//! You need to instantiate an [`Output`]
-//! for each output global you want to advertise to clients.
+//! You need to instantiate an [`Output`].
+//! To advertise a new output global to clients you then need to use [`Output::create_global`].
+//! The resulting `Global<WlOutput>` can later be destroyed again to stop advertising it
+//! without destroying it's state. E.g. in case the matching physical output got disabled at runtime.
 //!
-//! Just add it to your Display using the [`Output::new(..)`](Output::new)
-//! method. You can use the returned [`Output`] to change
+//! You can use the returned [`Output`] to change
 //! the properties of your output (if the current resolution mode changes for example),
 //! it'll automatically forward any changes to the clients.
+//!
+//! Additional protocols may piggy-back on this type.
+//! E.g. to also advertise an xdg-output for every wl-output you can use
+//! [`xdg::init_xdg_output_manager`].
+//!
+//! You can attach additional properties to your `Output`s by using [`Output::user_data`].
 //!
 //! ```
 //! # extern crate wayland_server;
@@ -23,8 +32,7 @@
 //!
 //! # let mut display = wayland_server::Display::new();
 //! // Create the Output with given name and physical properties
-//! let (output, _output_global) = Output::new(
-//!     &mut display,      // the display
+//! let output = Output::new(
 //!     "output-0".into(), // the name of this output,
 //!     PhysicalProperties {
 //!         size: (200, 150).into(),        // dimensions (width, height) in mm
@@ -34,6 +42,10 @@
 //!     },
 //!     None // insert a logger here
 //! );
+//! // create a global, if you want to advertise it to clients
+//! let _output_global = output.create_global(
+//!     &mut display,      // the display//!
+//! ); // you can drop the global, if you never intend to destroy it.
 //! // Now you can configure it
 //! output.change_current_state(
 //!     Some(Mode { size: (1920, 1080).into(), refresh: 60000 }), // the resolution mode,
@@ -102,7 +114,7 @@ pub struct PhysicalProperties {
 #[derive(Debug)]
 pub(crate) struct Inner {
     name: String,
-    pub(crate) log: ::slog::Logger,
+    description: String,
     instances: Vec<WlOutput>,
     physical: PhysicalProperties,
     location: Point<i32, Logical>,
@@ -112,7 +124,8 @@ pub(crate) struct Inner {
     current_mode: Option<Mode>,
     preferred_mode: Option<Mode>,
 
-    xdg_output: Option<XdgOutput>,
+    pub(crate) xdg_output: Option<XdgOutput>,
+    pub(crate) log: ::slog::Logger,
 }
 
 type InnerType = Arc<(Mutex<Inner>, UserDataMap)>;
@@ -164,7 +177,7 @@ impl Inner {
     }
 }
 
-/// An output as seen by the clients
+/// An abstract output.
 ///
 /// This handle is stored in the event loop, and allows you to notify clients
 /// about any change in the properties of this output.
@@ -174,17 +187,8 @@ pub struct Output {
 }
 
 impl Output {
-    /// Create a new output global with given name and physical properties
-    ///
-    /// The global is directly registered into the event loop, and this function
-    /// returns the state token allowing you to access it, as well as the global handle,
-    /// in case you wish to remove this global in the future.
-    pub fn new<L>(
-        display: &mut Display,
-        name: String,
-        physical: PhysicalProperties,
-        logger: L,
-    ) -> (Output, Global<WlOutput>)
+    /// Creates a new output state with the given name and physical properties.
+    pub fn new<L>(name: String, physical: PhysicalProperties, logger: L) -> Output
     where
         L: Into<Option<::slog::Logger>>,
     {
@@ -192,10 +196,12 @@ impl Output {
 
         info!(log, "Creating new wl_output"; "name" => &name);
 
+        let description = format!("{} - {} - {}", physical.make, physical.model, name);
+
         let inner = Arc::new((
             Mutex::new(Inner {
                 name,
-                log,
+                description,
                 instances: Vec::new(),
                 physical,
                 location: (0, 0).into(),
@@ -205,13 +211,25 @@ impl Output {
                 current_mode: None,
                 preferred_mode: None,
                 xdg_output: None,
+                log,
             }),
             UserDataMap::default(),
         ));
 
-        let output = Output { inner: inner.clone() };
+        Output { inner }
+    }
 
-        let global = display.create_global(
+    /// Create a new output global.
+    ///
+    /// The global is directly registered into the event loop, and this function
+    /// returns the global handle in case you wish to remove this global in the future.
+    ///
+    /// Calling this function multiple times without destroying the global in between,
+    /// will result in multiple globals, meaning the output will be advertised to clients
+    /// multiple times.
+    pub fn create_global(&self, display: &mut Display) -> Global<WlOutput> {
+        let inner = self.inner.clone();
+        display.create_global(
             3,
             Filter::new(move |(output, _version): (Main<WlOutput>, _), _, _| {
                 output.assign_destructor(Filter::new(|output: WlOutput, _, _| {
@@ -229,9 +247,7 @@ impl Output {
                 });
                 inner.0.lock().unwrap().new_global(output.deref().clone());
             }),
-        );
-
-        (output, global)
+        )
     }
 
     /// Attempt to retrieve a [`Output`] from an existing resource
@@ -269,6 +285,11 @@ impl Output {
         self.inner.0.lock().unwrap().current_mode
     }
 
+    /// Returns the preferred mode of the output
+    pub fn preferred_mode(&self) -> Option<Mode> {
+        self.inner.0.lock().unwrap().preferred_mode
+    }
+
     /// Returns the currently advertised transformation of the output
     pub fn current_transform(&self) -> Transform {
         self.inner.0.lock().unwrap().transform
@@ -289,9 +310,19 @@ impl Output {
         self.inner.0.lock().unwrap().name.clone()
     }
 
+    /// Returns the description of the output, if xdg-output is initialized
+    pub fn description(&self) -> String {
+        self.inner.0.lock().unwrap().description.clone()
+    }
+
     /// Returns the physical properties of the output
     pub fn physical_properties(&self) -> PhysicalProperties {
         self.inner.0.lock().unwrap().physical.clone()
+    }
+
+    /// Returns the currently advertised modes of the output
+    pub fn modes(&self) -> Vec<Mode> {
+        self.inner.0.lock().unwrap().modes.clone()
     }
 
     /// Removes a mode from the list of known modes